"Minor Text Fixes" and getting ready for upload
Removed Logger's OLD and INACTIVE token and put the token where it belongs into an ignored config.json file.

Removed the O from the Raid ID potentials.  People complained about it today, although 0 was already not an option.

Made Mewtwo raids (with poke.id == 150) last for 10 hours before expiring

altered how the embed looks for the raid event, so it now shows the username and the mention, which can be two different things.  Mentions are weird, because on mobile (android) they appear as the actual string.   GuildMember has a nickname, but it's hard to get to directly from the user, since you don't know what guild you're listing it in.