Two fixes: [Object object] and seperate config.json  files
There was an error when RaiderBot would load raids from the disk, the "owner" object wouldn't properly be set as a User, and so would sometimes improperly be printed as "raid.owner", intending for the toString() to be called, which is the mention.  I've updated the code so that all references now are just "raid.owner.mention"

Additionally, I took this opportunity to make a slight change to the loading of the config.json file.  This is a MAJOR CHANGE!   Now, Raider.js should be called with 1 argument, which is the name of the config file.  This way, a single install can run multiple versions of the bot.